+++ src/wenmode/plugins/__init__.py
@@ -1,21 +1,7 @@
 """Built-in Wenmode plugins and plugin API types."""
 
-from .._declarative import BlockFenced as BlockFenced
-from .._declarative import DeclarativePluginSpec as DeclarativePluginSpec
-from .._declarative import InlineDelimited as InlineDelimited
-from .._declarative import InlineLiteral as InlineLiteral
-from .._declarative import RendererFallback as RendererFallback
-from .._declarative import RenderTemplate as RenderTemplate
-from .._declarative import install_declarative as install_declarative
 from .types import RendererHandlers as RendererHandlers
 
 __all__ = [
-    'BlockFenced',
-    'DeclarativePluginSpec',
-    'InlineDelimited',
-    'InlineLiteral',
     'RendererHandlers',
-    'RendererFallback',
-    'RenderTemplate',
-    'install_declarative',
 ]