+++ tests/test_wenmode.py
@@ -6,20 +6,10 @@ from types import ModuleType
 import pytest
 
 from wenmode import Parser, Wenmode
+from wenmode._declaratives import BlockFenced, InlineDelimited, InlineLiteral
 from wenmode.directives import Admonition
 from wenmode.nodes import Literal, Parent
-from wenmode.plugins import (
-    BlockFenced,
-    DeclarativePluginSpec,
-    InlineLiteral,
-    RendererFallback,
-    RenderTemplate,
-    inline_math,
-    install_declarative,
-    mark,
-    ruby,
-    smartypants,
-)
+from wenmode.plugins import inline_math, mark, ruby, smartypants
 from wenmode.renderers import HTMLRenderer, RenderContext
 from wenmode.rules import AtxHeading, ContainerDirective, Link
 
@@ -112,7 +102,7 @@ def test_wenmode_installs_declarative_plugin() -> None:
     wen = Wenmode(plugins=[mark])
 
     assert mark.nodes == [mark.MarkNode]
-    assert mark.spec.syntax[0].opener == '=='
+    assert mark.rules[0].pattern == '=='
     assert wen.render('==marked *text*==\n') == '<p><mark>marked <em>text</em></mark></p>\n'
     assert wen.render('===not marked===\n') == '<p>===not marked===</p>\n'
 
@@ -127,26 +117,23 @@ def test_wenmode_installs_declarative_plugin_handlers() -> None:
 
 def test_wenmode_installs_declarative_inline_literal() -> None:
     class InlineLiteralPlugin:
-        spec = DeclarativePluginSpec(
-            name='custom_inline_literal',
-            syntax=[
-                InlineLiteral(
-                    name='custom_inline_literal',
-                    node=CustomInlineLiteral,
-                    opener='$',
-                    closer='$',
-                    reject_adjacent_delimiter=True,
-                    reject_closing_before_digit=True,
-                )
-            ],
-            renderers={'html': {CustomInlineLiteral.type: RendererFallback('value')}},
-        )
-
         nodes = [CustomInlineLiteral]
+        rules = [
+            InlineLiteral(
+                name='custom_inline_literal',
+                node=CustomInlineLiteral,
+                opener='$',
+                closer='$',
+                reject_adjacent_delimiter=True,
+                reject_closing_before_digit=True,
+            )
+        ]
+        handlers = {'html': {CustomInlineLiteral.type: lambda renderer, node, context: node.value}}
 
         @staticmethod
         def setup(wen: Wenmode, /) -> None:
-            install_declarative(wen, InlineLiteralPlugin.spec)
+            wen.register_rules(InlineLiteralPlugin.rules)
+            wen.register_renderer_handlers(InlineLiteralPlugin.handlers)
 
     wen = Wenmode(plugins=[InlineLiteralPlugin])
 
@@ -154,27 +141,84 @@ def test_wenmode_installs_declarative_inline_literal() -> None:
     assert wen.render('$value$5\n') == '<p>$value$5</p>\n'
 
 
-def test_wenmode_installs_declarative_block_fenced_literal() -> None:
-    class BlockLiteralPlugin:
-        spec = DeclarativePluginSpec(
-            name='custom_block_literal',
-            syntax=[
-                BlockFenced(
-                    name='custom_block_literal',
-                    node=CustomBlockLiteral,
-                    opener='%%%',
-                    closer='%%%',
-                    strip_content=True,
-                )
-            ],
-            renderers={'html': {CustomBlockLiteral.type: RenderTemplate('<aside>{value}</aside>\n')}},
-        )
-
-        nodes = [CustomBlockLiteral]
+def test_wenmode_installs_declarative_literal_value() -> None:
+    class DelimitedValuePlugin:
+        nodes = [CustomInlineLiteral]
+        rules = [
+            InlineLiteral(
+                name='custom_delimited_value',
+                node=CustomInlineLiteral,
+                opener='{%',
+                closer='%}',
+                reject_opening_whitespace=False,
+                reject_closing_whitespace=False,
+                reject_longer_run=False,
+            )
+        ]
+        handlers = {'html': {CustomInlineLiteral.type: lambda renderer, node, context: f'<data>{node.value}</data>'}}
 
         @staticmethod
         def setup(wen: Wenmode, /) -> None:
-            install_declarative(wen, BlockLiteralPlugin.spec)
+            wen.register_rules(DelimitedValuePlugin.rules)
+            wen.register_renderer_handlers(DelimitedValuePlugin.handlers)
+
+    wen = Wenmode(plugins=[DelimitedValuePlugin])
+
+    assert wen.render('A {% raw *value* %}.\n') == '<p>A <data> raw *value* </data>.</p>\n'
+
+
+def test_wenmode_installs_declarative_delimited_children_with_asymmetric_delimiters() -> None:
+    class DelimitedChildrenPlugin:
+        nodes = [CustomBlockParent]
+        rules = [
+            InlineDelimited(
+                name='custom_delimited_children',
+                node=CustomBlockParent,
+                opener='{+',
+                closer='+}',
+                reject_longer_run=False,
+            )
+        ]
+        handlers = {
+            'html': {
+                CustomBlockParent.type: lambda renderer, node, context: (
+                    f'<span>{renderer.render_children(node.children, context)}</span>'
+                )
+            }
+        }
+
+        @staticmethod
+        def setup(wen: Wenmode, /) -> None:
+            wen.register_rules(DelimitedChildrenPlugin.rules)
+            wen.register_renderer_handlers(DelimitedChildrenPlugin.handlers)
+
+    wen = Wenmode(plugins=[DelimitedChildrenPlugin])
+
+    assert wen.render('A {+very *important*+} value.\n') == (
+        '<p>A <span>very <em>important</em></span> value.</p>\n'
+    )
+
+
+def test_wenmode_installs_declarative_block_fenced_literal() -> None:
+    class BlockLiteralPlugin:
+        nodes = [CustomBlockLiteral]
+        rules = [
+            BlockFenced(
+                name='custom_block_literal',
+                node=CustomBlockLiteral,
+                opener='%%%',
+                closer='%%%',
+                strip_content=True,
+            )
+        ]
+        handlers = {
+            'html': {CustomBlockLiteral.type: lambda renderer, node, context: f'<aside>{node.value}</aside>\n'}
+        }
+
+        @staticmethod
+        def setup(wen: Wenmode, /) -> None:
+            wen.register_rules(BlockLiteralPlugin.rules)
+            wen.register_renderer_handlers(BlockLiteralPlugin.handlers)
 
     wen = Wenmode(plugins=[BlockLiteralPlugin])
 
@@ -183,25 +227,24 @@ def test_wenmode_installs_declarative_block_fenced_literal() -> None:
 
 def test_wenmode_installs_declarative_block_fenced_children() -> None:
     class BlockParentPlugin:
-        spec = DeclarativePluginSpec(
-            name='custom_block_parent',
-            syntax=[
-                BlockFenced(
-                    name='custom_block_parent',
-                    node=CustomBlockParent,
-                    opener='+++',
-                    closer='+++',
-                    content='children',
-                )
-            ],
-            renderers={'html': {CustomBlockParent.type: RendererFallback('children')}},
-        )
-
         nodes = [CustomBlockParent]
+        rules = [
+            BlockFenced(
+                name='custom_block_parent',
+                node=CustomBlockParent,
+                opener='+++',
+                closer='+++',
+                content='children',
+            )
+        ]
+        handlers = {
+            'html': {CustomBlockParent.type: lambda renderer, node, context: renderer.render_children(node.children, context)}
+        }
 
         @staticmethod
         def setup(wen: Wenmode, /) -> None:
-            install_declarative(wen, BlockParentPlugin.spec)
+            wen.register_rules(BlockParentPlugin.rules)
+            wen.register_renderer_handlers(BlockParentPlugin.handlers)
 
     wen = Wenmode(plugins=[BlockParentPlugin])
 
@@ -212,26 +255,25 @@ def test_declarative_block_fenced_children_respects_max_container_depth() -> Non
     openers = ['+++', '===', '%%%', '$$$', '???', '!!!', '&&&', ';;;']
 
     class BlockParentPlugin:
-        spec = DeclarativePluginSpec(
-            name='bounded_custom_block_parent',
-            syntax=[
-                BlockFenced(
-                    name=f'custom_block_parent_{index}',
-                    node=CustomBlockParent,
-                    opener=opener,
-                    closer=f'/{opener}',
-                    content='children',
-                )
-                for index, opener in enumerate(openers)
-            ],
-            renderers={'html': {CustomBlockParent.type: RendererFallback('children')}},
-        )
-
         nodes = [CustomBlockParent]
+        rules = [
+            BlockFenced(
+                name=f'custom_block_parent_{index}',
+                node=CustomBlockParent,
+                opener=opener,
+                closer=f'/{opener}',
+                content='children',
+            )
+            for index, opener in enumerate(openers)
+        ]
+        handlers = {
+            'html': {CustomBlockParent.type: lambda renderer, node, context: renderer.render_children(node.children, context)}
+        }
 
         @staticmethod
         def setup(wen: Wenmode, /) -> None:
-            install_declarative(wen, BlockParentPlugin.spec)
+            wen.register_rules(BlockParentPlugin.rules)
+            wen.register_renderer_handlers(BlockParentPlugin.handlers)
 
     markdown = (
         ''.join(f'{opener}\n' for opener in openers)
@@ -290,18 +332,19 @@ def test_wenmode_rejects_constructor_plugin_option_tuples() -> None:
         Wenmode(plugins=[(smartypants, {'dashes': False})])
 
 
-def test_wenmode_uses_setup_even_when_plugin_exposes_spec() -> None:
-    class SpecBackedPlugin:
-        spec = DeclarativePluginSpec(name='spec_backed', syntax=[], renderers={})
+def test_wenmode_uses_setup_even_when_plugin_exposes_metadata() -> None:
+    class MetadataPlugin:
+        name = 'metadata'
+        nodes: list[type[Parent]] = []
         called = False
 
         @staticmethod
         def setup(wen: Wenmode, /) -> None:
-            SpecBackedPlugin.called = True
+            MetadataPlugin.called = True
 
-    Wenmode().use(SpecBackedPlugin)
+    Wenmode().use(MetadataPlugin)
 
-    assert SpecBackedPlugin.called is True
+    assert MetadataPlugin.called is True
 
 
 def test_wenmode_registers_renderer_handlers_for_current_renderer() -> None: